pyproject.toml
[project]
name = "meisencam"
version = "0.1.0"
description = "Raspberry Pi bird camera with motion detection and Nextcloud upload"
requires-python = ">=3.11"
dependencies = [
    "picamera2; sys_platform == 'linux'",
    "requests",
    "Pillow",
]

[project.scripts]
meisencam = "meisencam.__main__:main"

[dependency-groups]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meisencam"]

[tool.uv]
# picamera2 depends on python-prctl which can only be built on Linux.
# Split environments so uv never tries to build linux-only sdists on macOS.
environments = ["sys_platform == 'linux'", "sys_platform != 'linux'"]
no-build-package = ["python-prctl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
